Extract remaining TextData functions: CONCATENATE(), JOIN() and REPT()
